requirements.txt
Brotli==1.0.9
certifi==2022.12.7
charset-normalizer==3.1.0
cloudpickle==2.2.1
cmake==3.26.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
decorator==4.4.2
dlib==19.21.0
ffmpeg-python==0.2.0
fonttools==4.39.3
future==0.18.3
hyperopt==0.2.7
idna==3.4
imageio==2.27.0
imageio-ffmpeg==0.4.8
importlib-resources==5.12.0
kiwisolver==1.4.4
matplotlib==3.7.1
moviepy==1.0.3
mutagen==1.46.0
networkx==3.1
numpy==1.22.0
opencv-python==4.6.0.66
packaging==23.1
pandas==2.0.0
Pillow==9.5.0
proglog==0.1.10
py4j==0.10.9.7
pycryptodomex==3.17
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
requests==2.28.2
scikit-image==0.17.2
scipy==1.10.1
six==1.16.0
sox==1.4.1
tifffile==2023.4.12
tqdm==4.65.0
tzdata==2023.3
urllib3==1.26.15
websockets==11.0.2
wget==3.2
wincertstore==0.2
xmltodict==0.13.0
yt-dlp==2023.3.4
zipp==3.15.0
